added tests, mediaMatch polyfill and separated App component from index.js #235
