+++ src/index.js
@@ -1,34 +1,8 @@
-import React from "react";
 import ReactDOM from "react-dom";
-import { RoutingApp } from "./modules/RoutingApp";
-import store from "./store";
-import { Provider } from "react-redux";
-import { STUY_SPEC_API_URL } from "./constants";
-
-import { ApolloProvider } from "react-apollo";
-import { ApolloClient } from "apollo-client";
-import { HttpLink } from "apollo-link-http";
-import { InMemoryCache } from "apollo-cache-inmemory";
-
-import { objectFilter } from "./utils";
-
-const apolloClient = new ApolloClient({
-  link: new HttpLink({ uri: `${STUY_SPEC_API_URL}/graphql` }),
-  cache: new InMemoryCache(),
-  connectToDevTools: true
-});
-
-Object.filter = objectFilter;
-
-const App = () => (
-  <Provider store={store}>
-    <ApolloProvider client={apolloClient}>
-      <RoutingApp />
-    </ApolloProvider>
-  </Provider>
-);
+import React from "react";
+import App from "./App";
 
 ReactDOM.render(
-  <App />,
-  document.getElementById("app"),
+	<App />,
+	document.getElementById("app"),
 );
