+++ src/setupTests.js
@@ -1,0 +1,7 @@
+window.matchMedia = window.matchMedia || function() {
+	return {
+		matches : false,
+		addListener : function() {},
+		removeListener: function() {}
+	};
+};
