+++ src/tests/App.spec.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from '../App';
+
+it('renders without crashing', () => {
+
+    const div = document.createElement('div');
+    ReactDOM.render(<App/>, div);
+    ReactDOM.unmountComponentAtNode(div);
+
+});
